requirements.txt
alembic==1.14.0
Babel==2.9.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Migrate==4.0.7
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.0.0
Mako==1.3.8
MarkupSafe==2.1.1
psycopg2-binary==2.9.10
python-dateutil==2.6.0
python-dotenv==1.0.1
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.34
typing_extensions==4.12.2
Werkzeug==2.0.3
WTForms==3.2.1
zipp==3.21.0
